improve error reporting
- print the group index which causes an error
- print the string which causes a parse error
- do not prefix error messages with 'error', as clap (sometimes) does
- validate each cpu in a range, not just the bounds
- disable color in clap output
- make error message formats more consistent

+++ src/util/format.rs
@@ -21,10 +21,10 @@ pub(crate) fn frequency(f: Frequency) -> String {
 }
 
 pub(crate) fn power(p: Power) -> String {
-    let mw = p.as_milliwatts().trunc();
-    if 0. == mw {
+    let mw = p.as_milliwatts().trunc() as u64;
+    if 0 == mw {
         "0 W".to_string()
-    } else if mw < 1000. || mw % 1000. == 0. {
+    } else if mw < 1000 || mw % 1000 == 0 {
         format!("{:.0}", p)
     } else {
         format!("{:.1}", p)
